requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
redis==5.0.1
httpx==0.25.2
pydantic==2.5.0
pydantic-settings==2.1.0
apscheduler==3.10.4
anthropic==0.7.8
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
openpyxl==3.1.2
python-telegram-bot==20.7
asyncpg==0.29.0
python-dateutil==2.8.2
cryptography==41.0.8
ed25519==1.5
